changed file size return value to f64. small QOL on frontend for display values

+++ src-tauri/src/models/mod.rs
@@ -34,7 +34,7 @@ pub struct ExtractedTrack {
     pub album: String,
     pub duration_ms: i64,
     pub file_format: String, // ex. "mp3"
-    pub file_size: i64,
+    pub file_size: f64,
     pub date_added: i64, // ex. 20260112
     #[serde(skip_serializing_if = "Option::is_none")]
     pub thumbnail_base64: Option<String>,
